refactor(agent): author the containerd config as TOML, not a Rust string
Escaped quotes, `\n\` continuations and `\x20` to get a leading space
were the sign that a string literal was fighting the language it
contained. The config is now a real `.toml` file with `@NAME@`
placeholders, `include_str!`d and substituted from
`arcbox_constants::paths` — so the CNI paths stay the same values the
agent uses elsewhere to create those directories, rather than being
copied.

`include_str!` rather than a file in the guest rootfs, deliberately. The
rootfs is a separate release train, so a config living there could lag
the agent that depends on it, and an option would go missing silently —
which is exactly the failure mode the overlayfs section is most exposed
to. The rootfs is also read-only, so the agent has to write this at
runtime either way; moving only the *content* across a repo boundary
buys nothing.

It moves out of `agent/linux/` at the same time, for the reason already
noted on `live_exports`: everything in that module is
`cfg(target_os = "linux")`, so its tests never compile on a macOS dev
host, and CI excludes this crate entirely. Rendering a config file is
pure string work, so it belongs where the tests can run — and they do:
four of them, including one for the failure mode a template introduces.

That one earned its place immediately. `mount_options = @FOO@` would at
least be a parse error, but `bin_dir = "@FOO@"` is valid TOML carrying a
nonsense path, which nothing downstream would report. It failed on its
first run — on the sentence in the file's own header explaining the
placeholder convention — so it now exempts comments.
